requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.26
pydantic-settings>=2.2
feedparser>=6.0
cachetools>=5.3

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pgvector>=0.2

# Authentication
fastapi-users[sqlalchemy]>=13.0

# OAuth
httpx-oauth>=0.14

# Security
cryptography>=42.0

# Rate Limiting
slowapi>=0.1.9
